requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
influxdb-client==1.42.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
PyYAML==6.0.1
pyzmq==26.0.2
reactivex==4.0.4
regex==2024.4.28
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.6.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
